emr_platform/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
psycopg2-binary
web3
python-dotenv
